Consolidate math, cursor, and event helpers into single utils.js
