--- src/helpers/math.js
@@ -1,51 +1,0 @@
-/**
- * @param {number} x
- * @param {number} y
- * @returns {number}
- */
-export function hypot(x, y) {
-  return Math.sqrt(x * x + y * y)
-}
-
-/**
- * @param {number} num
- * @param {number} cap
- * @returns {number | any}
- */
-export function min(num, cap) {
-  let neg = cap * -1
-  return num > neg && num < 0 ? neg : num < cap && num > 0 ? cap : num
-}
-
-/**
- * @param {number} num
- * @param {number} cap
- * @returns {any | number}
- */
-export function max(num, cap) {
-  let neg = cap * -1
-  return num > cap ? cap : num < neg ? neg : num
-}
-
-/**
- * @param {number} value
- * @param {number} min
- * @param {number} max
- * @returns {number}
- */
-export function clamp(value, min, max) {
-  return Math.min(Math.max(value, min), max)
-}
-
-/**
- * @param {number} x
- * @param {number} y
- * @returns {number}
- */
-export function angle(x, y) {
-  let angle = Math.atan2(y, x) * (180 / Math.PI)
-  if (angle < 0) {
-    angle += 360
-  }
-  return angle
-}
